arch: arm: boot: dts: do not override model in Maia SDR dts
The Pluto u-boot checks the device tree model to detect whether the
Pluto is rev. C or not, so if we override the model we make u-boot
believe that the Pluto is never rev. C.

Additionally, the model was currently not being overridden correctly
and all the device trees ended up saying "rev. B".

Signed-off-by: Daniel Estévez <daniel@destevez.net>